requirements.txt
absl-py==1.0.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
appnope==0.1.3
argcomplete==1.10.3
argilla==1.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
backcall==0.2.0
backoff==2.2.1
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.8.2
bleach==5.0.1
blinker==1.5
bs4==0.0.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
chromadb==0.3.21
ci-info==0.3.0
click==8.0.3
clickhouse-connect==0.5.20
cloudpickle==2.0.0
commonmark==0.9.1
compressed-rtf==1.0.6
configobj==5.0.8
configparser==5.3.0
cryptography==36.0.0
cycler==0.11.0
d2l==0.16.0
dataclasses-json==0.5.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
dm-tree==0.1.7
docx2txt==0.8
duckdb==0.7.1
ebcdic==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
etelemetry==0.3.0
executing==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.3
fastapi==0.95.1
fastjsonschema==2.16.2
filelock==3.4.2
fitz==0.0.1.dev2
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.5
frontend==0.0.3
frozenlist==1.3.3
future==0.18.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpt-index==0.5.15
gptcache==0.1.11
greenlet==2.0.2
grpcio==1.43.0
h11==0.14.0
h5py==3.6.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==2.10
IMAPClient==2.1.0
implicit==0.6.2
importlib-metadata==4.10.0
iniconfig==1.1.1
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterplot==0.0.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
langchain==0.0.139
libclang==12.0.0
llama-index==0.5.15
llvmlite==0.37.0
looseversion==1.1.2
lrcurve==1.1.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.6.1
nest-asyncio==1.5.6
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
notebook==6.4.12
numba==0.54.1
numpy==1.22.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.1
odfpy==1.4.1
olefile==0.46
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.0
pandas==2.0.0
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.1
plotly==5.10.0
pluggy==0.13.1
posthog==2.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.1
prov==2.0.0
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydeck==0.8.0b3
pydot==1.4.2
Pygments==2.13.0
PyGObject==3.38.0
Pympler==1.0.1
pyOpenSSL==21.0.0
pypandoc==1.11
pyparsing==2.4.7
pypdf==3.8.1
PyPDF2==3.0.1
PyQt5==5.15.2
PyQt5_sip==4.19.24
pyrsistent==0.18.1
pytest==6.2.4
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.5
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rdflib==6.3.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==12.5.1
rsa==4.8
scikit-learn==1.0.1
scikit-learn-extra==0.3.0
scikit-surprise==1.1.3
scipy==1.7.1
selenium==4.1.0
semver==2.13.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shap==0.40.0
simplejson==3.19.1
six==1.12.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sparse==0.14.0
SpeechRecognition==3.8.1
SQLAlchemy==1.4.47
stack-data==0.5.1
starlette==0.26.1
streamlit==1.13.0
streamlit-chat==0.0.2.2
streamlit-lottie==0.0.3
surprise==0.1
TBB==0.1
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-probability==0.16.0
termcolor==1.1.0
terminado==0.16.0
textract==1.6.5
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.0.0
tiktoken==0.3.3
tinycss2==1.1.1
tk==0.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.62.3
traitlets==5.4.0
traits==6.3.2
transformers==4.27.4
trio==0.19.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.2
tzlocal==4.2
unstructured==0.5.12
urllib3==1.26.7
uvicorn==0.21.1
uvloop==0.17.0
validators==0.20.0
virtualenv==20.10.0
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==11.0.1
Werkzeug==2.0.2
widgetsnbextension==4.0.3
wrapt==1.13.3
wsproto==1.0.0
xgboost==1.5.0
xlrd==1.2.0
XlsxWriter==3.1.0
yarl==1.8.2
zipp==3.7.0
zstandard==0.20.0
